pyproject.toml
[project]
name = "oraclewatch"
version = "0.1.0"
description = "Real-time monitoring and alerting for new prediction markets on Kalshi and Polymarket"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }

dependencies = [
    "httpx>=0.27,<1",
    "python-dotenv>=1.0,<2",
    "aiosqlite>=0.20,<1",
    "pydantic>=2.5,<3",
    "rich>=13.0,<14",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
oraclewatch = "src.main:cli_entry"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
